requirements.txt
streamlit==1.32.0
numpy==1.26.4
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.19.0